Restructure src dir
