+++ packages/experimental-app-router/src/index.tsx
@@ -1,3 +1,3 @@
 export { getClient, getAuthClient } from './client.js';
-export { nextRouteHandler } from './server/nextRouteHandler.js';
-export { fetchAccessToken } from './server/fetchAccessToken.js';
+export { faustRouteHandler } from './server/routeHandler/index.js';
+export { fetchAccessToken } from './server/auth/fetchAccessToken.js';
